랜딩페이지 공통컴포넌트 생성
